refactor: src/test/java/seedu/address/model/
